Move parameter functions to workflow.py, add inline comments
