+++ src/wei/core/workflow.py
@@ -1,5 +1,7 @@
 """The module that initializes and runs the step by step WEI workflow"""
 
+import copy
+import re
 from typing import Any, Dict, List, Optional
 
 from fastapi import UploadFile
@@ -143,3 +145,97 @@ def cancel_active_workflow_runs() -> None:
             WorkflowStatus.IN_PROGRESS,
         ]:
             cancel_workflow_run(wf_run)
+
+
+def insert_parameter_values(workflow: Workflow, parameters: Dict[str, Any]):
+    """Replace the parameter strings in the workflow with the provided values"""
+    for param in workflow.parameters:
+        if param.name not in parameters.keys():
+            if param.default:
+                parameters[param.name] = param.default
+            else:
+                raise ValueError(
+                    "Workflow parameter: "
+                    + param.name
+                    + " not provided, and no default value is defined."
+                )
+    steps = []
+    for step in workflow.flowdef:
+        for key, val in iter(step):
+            if type(val) is str:
+                setattr(step, key, value_substitution(val, parameters))
+
+        step.args = walk_and_replace(step.args, parameters)
+        steps.append(step)
+    workflow.flowdef = steps
+
+
+def walk_and_replace(args: Dict[str, Any], input_parameters: Dict[str, Any]):
+    """Recursively walk the arguments and replace all parameters"""
+    new_args = copy.deepcopy(args)
+    for key in args.keys():
+        if type(args[key]) is str:
+            new_args[key] = value_substitution(args[key], input_parameters)
+        elif type(args[key]) is dict:
+            new_args[key] = walk_and_replace(args[key], input_parameters)
+        if type(key) is str:
+            new_key = value_substitution(key, input_parameters)
+            new_args[new_key] = new_args[key]
+            if key is not new_key:
+                new_args.pop(key, None)
+    return new_args
+
+
+def value_substitution(input_string: str, input_parameters: Dict[str, Any]):
+    """Perform $-string substitution on input string, returns string with substituted values"""
+    # * Check if the string is a simple parameter reference
+    if type(input_string) is str and re.match(r"^\$[A-z0-1_\-]*$", input_string):
+        if input_string.strip("$") in input_parameters.keys():
+            input_string = input_parameters[input_string.strip("$")]
+        else:
+            raise ValueError(
+                "Unknown parameter:"
+                + input_string
+                + ", please define it in the parameters section of the Workflow Definition."
+            )
+    else:
+        # * Replace all parameter references contained in the string
+        working_string = input_string
+        for match in re.findall(
+            r"((?<!\$)\$(?!\$)[A-z0-1_\-\{]*)(\s|\}|$)", input_string
+        ):
+            param_name = match[0].strip("$")
+            param_name = param_name.strip("{")
+            if param_name in input_parameters.keys():
+                if match[1] == "}":
+                    if match[0][1] == "{":
+                        working_string = re.sub(
+                            r"((?<!\$)\$(?!\$)[A-z0-1_\-\{]*)(\})",
+                            str(input_parameters[param_name]),
+                            working_string,
+                        )
+                        input_string = working_string
+                    else:
+                        raise SyntaxError(
+                            "forgot opening { in parameter insertion: " + match[0] + "}"
+                        )
+                elif match[1] == " ":
+                    working_string = re.sub(
+                        r"((?<!\$)\$(?!\$)[A-z0-1_\-\{]*)(\s)",
+                        str(input_parameters[param_name]) + " ",
+                        working_string,
+                    )
+                elif match[1] == "":
+                    working_string = re.sub(
+                        r"((?<!\$)\$(?!\$)[A-z0-1_\-\{]*)($)",
+                        str(input_parameters[param_name]),
+                        working_string,
+                    )
+                input_string = working_string
+            else:
+                raise ValueError(
+                    "Unknown parameter:"
+                    + param_name
+                    + ", please define it in the parameters section of the Workflow Definition."
+                )
+    return input_string
